Snap.Predicates.Params

Documentation

data Param

A Predicate looking for some parameter.

Constructors

Param ByteString

data ParamTrans a

A Predicate looking for some parameter, and returning as TVal the result of applying the given function to all parameter values.

Constructors

ParamTrans ByteString ([ByteString] -> a)